Renamed files to match dart syntax
